+++ src/marked.js
@@ -71,7 +71,7 @@ marked.Tokenizer = Tokenizer;
 marked.Slugger = Slugger;
 marked.Hooks = Hooks;
 marked.parse = marked;
-marked.instance = Marked;
+marked.Instance = Marked;
 
 export const options = marked.options;
 export const setOptions = marked.setOptions;
