+++ test/unit/utils.js
@@ -1,0 +1,5 @@
+export async function timeout(ms = 1) {
+  return new Promise(resolve => {
+    setTimeout(resolve, ms);
+  });
+}
